Add files via upload
These are the important packets for metasploit. The important thing to note is the large amount of NT Trans Requests and the one packet with malformed packet at the end(highlighted). The large payload of trans requests leads to the secondary NT Trans Requests that are malformed, acting as sort of trigger point for the vulnerability and launcher for the remote malware on the attacker machine.